Patch query_method until Rails4
